refactor: Change TcpMavConnection to AbstractMavConnection and move it to connection-core
